Fix bundle path for standalone (#395)
**Story card:** [ch6351](https://app.shortcut.com/simpledotorg/story/6351/setup-dual-dependency-approach-to-prepare-for-rails-6-and-beyond)

## Because

Bootboot has a tough time in our standalone deployment environment because it configure's bundler to look for our Gemfile in an absolute release path. This release path changes with every release, of course. 

## This addresses

Updates bundler's config to look for the Gemfile in the more permanent "shared" directory of ansistrano.

Note: Ansistrano tacks on a "Z" to our release directories (lol), so this sed expression is _slightly_ different from our [capistrano fix](https://github.com/simpledotorg/simple-server/pull/3270).

This PR is a follow-up to
* https://github.com/simpledotorg/simple-server/pull/3270
* https://github.com/simpledotorg/simple-server/pull/3272